Suppressed UserBoards class. Use BoardsManager instead.
